+++ mapmatching/match/dir_similarity.py
@@ -1,21 +1,15 @@
 from geopandas import GeoDataFrame
 
-from .misc import get_shared_line, merge_step_arrs
 from ..geo.azimuth import cal_linestring_azimuth_cos_dist
-
+from .candidatesGraph import get_shortest_geometry
 
 def cal_dir_prob(gt:GeoDataFrame, geom='geometry'):
-    # TODO: 适用 sub_string 代替现有的情况
+    # TODO: 使用 sub_string 代替现有的情况
     # Add: dir_prob
     def _cal_dir_similarity(x):
         return cal_linestring_azimuth_cos_dist(x[geom], x['move_dir'], weight=True)
-
-    # complete shortest path
-    mask = gt.flag == 1
-    gt.loc[mask, geom] = gt.loc[mask].apply(lambda x: 
-        get_shared_line(x.step_0, x.step_n), axis=1)
-    gt.loc[~mask, geom] = gt.loc[~mask].apply(
-        merge_step_arrs, axis=1)
+    
+    gt = get_shortest_geometry(gt, geom, format='array')
     gt.loc[:, 'dir_prob'] = gt.apply(_cal_dir_similarity, axis=1)
     
     return gt