+++ mapmatching/matching.py
@@ -11,7 +11,6 @@ from .graph import GeoDigraph
 from .update_network import check_steps
 from .geo.metric import lcss, edr, erp
 from .geo.ops import check_duplicate_points
-from .geo.ops.point2line import project_points_2_linestrings
 from .geo.ops.simplify import simplify_trajetory_points
 from .geo.ops.resample import resample_polyline_seq_to_point_seq, resample_point_seq
 
@@ -19,18 +18,19 @@ from .osmnet.build_graph import build_geograph
 
 from .match.status import STATUS
 from .match.io import load_points
-from .match.postprocess import get_path
-from .match.candidatesGraph import construct_graph
 from .match.spatialAnalysis import analyse_spatial_info
 from .match.geometricAnalysis import analyse_geometric_info
+from .match.postprocess import get_path, project, transform_mathching_res_2_path
+from .match.candidatesGraph import construct_graph, get_shortest_geometry
 from .match.viterbi import process_viterbi_pipeline, find_matched_sequence
-from .match.visualization import plot_matching_result, debug_gt_level_parallel
+from .match.visualization import plot_matching_result, debug_traj_matching
 
 from .utils.timer import timeit
 from .utils.logger_helper import make_logger
 from .utils.misc import SET_PANDAS_LOG_FORMET
 from .setting import DATA_FOLDER, DEBUG_FOLDER
 
+from typing import List
 
 class ST_Matching():
     def __init__(self,
@@ -72,20 +72,22 @@ class ST_Matching():
     def matching(self, traj, top_k=None, dir_trans=False, beam_search=True,
                  simplify=True, tolerance=5, plot=False, save_fn=None,
                  debug_in_levels=False, details=False, metric=None, 
-                 check_duplicate=False, check_topo=False):
+                 check_duplicate=False, check_topo=False, search_radius=None):
         self.logger.trace("start")
         res = {'status': STATUS.UNKNOWN, 'ori_crs': deepcopy(traj.crs.to_epsg())}
 
-        _traj = self.align_crs(traj.copy())
         # simplify trajectory
+        _traj = self.align_crs(traj.copy())
         if simplify:
-            _traj = self.simplify(_traj, tolerance=tolerance) # tolerance, 5 meters
+            _traj = self.simplify(_traj, tolerance = tolerance) # tolerance, 5 meters
         elif check_duplicate:
             _traj = check_duplicate_points(_traj)
             
         # geometric analysis
         top_k = top_k if top_k is not None else self.top_k_candidates
-        cands = analyse_geometric_info(_traj, self.base_edges, top_k, self.cand_search_radius)
+        if search_radius is None:
+            search_radius = self.cand_search_radius
+        cands = analyse_geometric_info(_traj, self.base_edges, top_k, search_radius)
         
         # is_valid
         s, _ = self._is_valid_cands(_traj, cands, res)
@@ -96,15 +98,17 @@ class ST_Matching():
         res['probs'] = {}
         rList, graph = self.spatial_analysis(_traj, cands, dir_trans, beam_search, metric=res['probs'])
         match_res, steps = get_path(rList, graph, cands, metric=res['probs'])
+        
         if 'status' in res['probs']:
             res['status'] = res['probs']['status']
             del res['probs']['status']
         res.update(match_res)
 
+        # add details
         if details or check_topo:
             attrs = ['pid_1', 'step_0_len', 'step_n_len', 'cost', 'd_sht', 'd_euc', 'dist_prob', 
                      'trans_prob', 'observ_prob', 'prob', 'flag', 'status', 'dst', 'src','step_0', 
-                     'geometry', 'step_n', 'path', 'epath', 'vpath','dist']
+                     'geometry', 'step_n', 'path', 'epath', 'vpath', 'dist', 'dist_0', 'step_1']
             attrs = [i for i in attrs if i in list(graph)]
             if 'move_dir' in graph:
                 attrs += ['move_dir']
@@ -119,10 +123,12 @@ class ST_Matching():
             }
             res['details'] = _dict
 
+        # metric
         if metric is not None:
             res['metric'] = self.eval(_traj, res, metric=metric)
             self.logger.info(f"{metric}: {res['metric']}")
 
+        # plot
         if plot or save_fn:
             fig, ax = self.plot_result(traj, res)
             if plot:
@@ -132,11 +138,12 @@ class ST_Matching():
             if save_fn:
                 fig.savefig(save_fn, dpi=300, bbox_inches='tight', pad_inches=0.02)
 
+        # debug helper
         if debug_in_levels:
-            self.matching_debug(_traj, graph)
-        
+            self.debug_matching_among_levels(_traj, graph)
+
+        # check topo        
         if check_topo:
-            # TODO
             flag = check_steps(self, res, prob_thred=.75, factor=1.2)
             if flag:
                 res = self.matching(_traj, top_k, dir_trans, beam_search,
@@ -215,26 +222,10 @@ class ST_Matching():
         return _eval[0](*_eval[1])
 
     def project(self, points, path, keep_attrs=['eid', 'proj_point'], normalized=True, reset_geom=True):
-        _points = points[[points.geometry.name]]
-        ps = project_points_2_linestrings(
-            _points, 
-            path.to_crs(points.crs), 
-            normalized = normalized)
-        
-        if keep_attrs:
-            ps = ps[keep_attrs]
-
-        ps = gpd.GeoDataFrame(pd.concat([_points, ps], axis=1), crs=points.crs)
-        if reset_geom:
-            ps.loc[:, 'ori_geom'] = points.geometry.apply(lambda x: x.wkt)
-            ps.set_geometry('proj_point', inplace=True)
-            ps.drop(columns=['geometry'], inplace=True)
-        
-        return ps
+        return project(points, path, keep_attrs, normalized, reset_geom)
 
     def load_points(self, fn, simplify=False, tolerance: int = 10,
                     crs: int = None, in_sys: str = 'wgs', out_sys: str = 'wgs'):
-        
         traj, _ = load_points(fn, simplify, tolerance, crs, in_sys, out_sys)
         
         return traj
@@ -242,88 +233,26 @@ class ST_Matching():
     def simplify(self, points:gpd.GeoDataFrame, tolerance:int=None, inplace=False):        
         return simplify_trajetory_points(points, tolerance, inplace=True, logger=self.logger)
 
-    def matching_debug(self, traj:gpd.GeoDataFrame, graph:gpd.GeoDataFrame, level:int=None, debug_folder:str='./debug'):
-        """_summary_
+    def debug_matching_among_levels(self, traj: gpd.GeoDataFrame, graph: gpd.GeoDataFrame, 
+                                    level: list=None, debug_folder: str='./debug'):
 
-        Args:
-            traj (gpd.GeoDataFrame): _description_
-            graph (gpd.GeoDataFrame): _description_
-            level (int, optional): _description_. Defaults to None, namely output all layers.
-            debug_folder (str, optional): _description_. Defaults to './debug'.
-        """
-        graph = gpd.GeoDataFrame(graph)
-        
-        if level is None:
-            layer_ids = graph.index.get_level_values(0).unique().sort_values().values
-        else:
-            layer_ids = level if isinstance(level, list) else [level]
+        return debug_traj_matching(traj, graph, self.net, level, debug_folder)
 
-        for idx in layer_ids:
-            img = debug_gt_level_parallel(self.net, traj, graph, idx)
-            img.save(os.path.join(debug_folder, f"level_{idx}.jpg"))
-        
-        return img
-
-    @timeit
     def plot_result(self, traj, info):
         info = deepcopy(info)
         if info['status'] == 3:
             path = None
-        elif info.get('details', {}).get('path', None):
+        elif info.get('details', {}).get('path', None) is not None:
             path = info['details']['path']
         else:
             path = self.transform_res_2_path(info)
         
         fig, ax = plot_matching_result(traj, path, self.net)
-        if not info:
-            return fig, ax
-
-        for att in ['epath', "step_0", "step_n", 'details']:
-            if att not in info:
-                continue
-            info.pop(att)
-
-        text = []
-        if "probs" in info:
-            probs = info.pop('probs')
-            info.update(probs)
-        
-        for key, val in info.items():
-            if 'prob' in key:
-                _str = f"{key}: {val * 100: .2f} %"
-            else:
-                _str = f"{key}: {val}"
-            text.append(_str)
-
-        x0, x1, y0, y1 = ax.axis()
-        ax.text(x0 + (x1- x0)/50, y0 + (y1 - y0)/50, "\n".join(text))
 
         return fig, ax
 
     def transform_res_2_path(self, res, ori_crs=True, attrs=None):
-        if attrs is None:
-            attrs = ['eid', 'way_id', 'src', 'dst', 'name', 'road_type', 'link', 'speed', 'dist', 'geometry']
-        path = self.net.get_edge(res['epath'], attrs, reset_index=True)
-
-        _len = len(res['epath']) 
-        if _len == 1:
-            path.loc[0, 'dist'] *= res['step_n'] - res['step_0']
-            path.loc[0, 'geometry'] = shapely.ops.substring(
-                path.iloc[0].geometry, res['step_0'], res['step_n'], normalized=True)
-        else:
-            path.loc[0, 'dist'] *= 1 - res['step_0']
-            path.loc[0, 'geometry'] = shapely.ops.substring(
-                path.iloc[0].geometry, res['step_0'], 1, normalized=True)
-
-            path.loc[_len - 1, 'dist'] *= res['step_n']
-            path.loc[_len - 1, 'geometry'] = shapely.ops.substring(
-                path.iloc[-1].geometry, 0, res['step_n'], normalized=True)
-        
-        path = path[~path.geometry.is_empty]
-        if ori_crs:
-            path = path.to_crs(res['ori_crs'])
-
-        return path
+        return transform_mathching_res_2_path(res, self.net, ori_crs, attrs)
 
     def update(self):
         return NotImplementedError